install build system dependencies before recursing
When we see a build tool enter the set of packages being
processed, there is a good chance we will see circular
dependencies with other things also being packaged with that
tool, including 2nd or 3rd order dependencies of the tool
itself. For example, plugins to the tool often depend on the
tool. The code for looking up the metadata of a project can
cope with `backend-path` being set, but that does not help
*later* dependencies, so we install the tool to ensure it is
present when required. This introduces an undesirable
dependency on having these build tools in the system
already, so we may end up wanting to treat them as special