handle dists where the source dir name and dist name do not match
Dists like poetry-core end up in directories with names starting with
'poetry_core' because of the name transformation requirements of the
various packaging PEPs. Convert the dist name to the filename and look
for it inside the various stages of the build instead of trying to
pass the name via wildcard to the program from the outside.